Make the following changes:
,

Find this element:

  <div class="items-center p-8 pt-0 flex justify-center">

with the text content:

    Explore Biotech Careers

Relevant files:

And change it as follows:

create the careers section which is actually a guide and tips and tricks page for every beginners. It will have infographical representations about what are the different way the biotech subject will opens the doors in future from nanotech to ai in medical to every subject it connected with like genetic, gene design, drug discovery, to all the field. make a kind of family tree kind of schematic representations. Do the research and make the page accordingly. make it informative and easy to understand by the beginners.